add update and get and delete blog by id

+++ app/utils/function.js
@@ -69,8 +69,15 @@ function VerifyRefreshToken(token) {
 };
 
 function deleteFileInPublic(fileAddress) {
-    const pathFile = path.join(__dirname, "..","..","public",fileAddress);
-    fs.unlinkSync(pathFile);
+    if(fileAddress) {
+        const pathFile = path.join(__dirname, "..","..","public",fileAddress);
+        if(fs.existsSync(pathFile)) {
+            fs.unlinkSync(pathFile);
+        }
+        else {
+            return false;
+        }
+    }
 }
